Fixed: TRACK-113: Using realistic seed data only above average students are shown in performance data
There was a problem where the number of lessons for a student was incorrectly calculated. This has been fixed and the chart seems to work fine now.
Because of the heavy performance impact of this chart, I have made a selection on Group obliged for this chart. If no Group was selected, the chart is restricted to the first Group in the selection.